Changed the implementation of the bootstrap filter to be in pure python instead of using templates
I feel that using filters in templates in filters in templates might
be detrimental be a bad idea in term of design and performance.

I reimplemented the same behaviour in a python function. It seems to
me the logic is more clear that way.

Bonus : much less whitespaces in the output.